+++ src/main/java/org/dmdev/natalliavasilyeva/persistence/jpa/Price.java
@@ -1,43 +1,44 @@
 package org.dmdev.natalliavasilyeva.persistence.jpa;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class Price implements Entity, Serializable {
 
-    private int id;
-    private int sum;
+    private long id;
+    private BigDecimal sum;
 
     public Price() {
     }
 
-    private Price(int id, int sum) {
+    private Price(long id, BigDecimal sum) {
         this.id = id;
         this.sum = sum;
     }
 
     @Override
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public int getSum() {
+    public BigDecimal getSum() {
         return sum;
     }
 
-    public void setSum(int sum) {
+    public void setSum(BigDecimal sum) {
         this.sum = sum;
     }
 
     public static final class Builder {
-        private int id;
-        private int sum;
+        private long id;
+        private BigDecimal sum;
 
-        public Builder id(int id) {
+        public Builder id(long id) {
             this.id = id;
             return this;
         }
 
-        public Builder sum(int sum) {
+        public Builder sum(BigDecimal sum) {
             this.sum = sum;
             return this;
         }
